requirements.txt
requests
pandas
numpy
matplotlib
pillow